Content type matching problem with LinkedIn
I've recently encountered a problem using the LinkedIn configuration for this plugin. The rule for matching the content type should look like this:

m!^(application/json|text/javascript)(;\s*charset=\S+)?$!

because in the LinkedIn case there is no space between the content type and the charset definition. Also maybe you should include an else statement for this rule in case is not matched.